Consolidate Gameplay Level
Added a LevelData object that holds the building blocks for a level. Added prefab(s) for cell placement, and starting tile positions. Added a gameplay container object to "LevelOneRoundOne" scene which holds the objects created by LevelData. When the level is complete -> all children of the container are destroyed, and the next level's data is used to create new objects.

Known issue:
During testing of this new flow, the "client" player's round scores would show up as 0 after the first round - and give inaccurate information on the score board. The host scoreboard has the correct round scores and total scores... so the accurate information is being received. - Unsure what's causing the client-scoreboard to desync/display incorrectly.