Chat system upload and beginning of board screenshot
Initial upload of tabbed chat system for the three questions: "What Happened? "So What?" and "Now What?"

uploaded a script from a tutorial to take a screenshot and save it to a file.

 commented out some Debug.Log() lines that were triggered by a key press to avoid triggering in error while typing into a chat box.

ScoreBoardMatrix now takes two additional parameter images to update display (NYI).

ToDo: rework screenshot script to take a picture and send it over the network.